+++ packages/babel-preset-gatsby/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.17](https://github.com/gatsbyjs/gatsby/compare/babel-preset-gatsby@0.2.15...babel-preset-gatsby@0.2.17) (2019-09-26)
+
+**Note:** Version bump only for package babel-preset-gatsby
+
+## [0.2.16](https://github.com/gatsbyjs/gatsby/compare/babel-preset-gatsby@0.2.15...babel-preset-gatsby@0.2.16) (2019-09-26)
+
+**Note:** Version bump only for package babel-preset-gatsby
+
+## [0.2.15](https://github.com/gatsbyjs/gatsby/compare/babel-preset-gatsby@0.2.14...babel-preset-gatsby@0.2.15) (2019-09-25)
+
+**Note:** Version bump only for package babel-preset-gatsby
+
+## [0.2.14](https://github.com/gatsbyjs/gatsby/compare/babel-preset-gatsby@0.2.13...babel-preset-gatsby@0.2.14) (2019-09-20)
+
+**Note:** Version bump only for package babel-preset-gatsby
+
 ## [0.2.13](https://github.com/gatsbyjs/gatsby/compare/babel-preset-gatsby@0.2.12...babel-preset-gatsby@0.2.13) (2019-09-09)
 
 **Note:** Version bump only for package babel-preset-gatsby